+++ power-board.php
@@ -10,15 +10,15 @@
  * Description: PowerBoard is a payment orchestration solution that delivers payment choice, operational efficiency and security for your business.
  * Author: PowerBoard
  * Author URI: https://www.commbank.com.au/business/payments/take-online-payments/powerboard.html#getting-started
- * Version: 1.3.1
+ * Version: 1.4.0
  * Requires at least: 6.6
  * Text Domain: power-board
- * Tested up to: 6.7.1
- * Stable tag: 1.3.1
+ * Tested up to: 6.8
+ * Stable tag: 1.4.0
  * License: GPL-3.0
  * License URI: http://www.gnu.org/licenses/gpl-3.0.html
  * WC requires at least: 6.6
- * WC tested up to: 9.5.2
+ * WC tested up to: 9.8.1
  * Requires Plugins: woocommerce
  *
  * @noinspection PhpUndefinedFunctionInspection for plugin_dir_url
@@ -39,7 +39,7 @@ if ( ! defined( 'POWER_BOARD_PLUGIN_URL' ) ) {
 }
 
 if ( ! defined( 'POWER_BOARD_PLUGIN_VERSION' ) ) {
-	define( 'POWER_BOARD_PLUGIN_VERSION', '1.3.1' );
+	define( 'POWER_BOARD_PLUGIN_VERSION', '1.4.0' );
 }
 
 if ( ! defined( 'POWER_BOARD_PLUGIN_PREFIX' ) ) {