+++ templates/admin/admin.php
@@ -10,6 +10,11 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
+/* @noinspection PhpUndefinedFunctionInspection */
+\settings_fields( 'powerboard_settings' );
+/* @noinspection PhpUndefinedFunctionInspection */
+\settings_errors( 'powerboard_checkout_version' );
+
 if ( isset( $data['template_service'] ) ) {
 	$template_service = $data['template_service'];