+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [1.4.0] - 2025-06-02
+
+### Compatibility
+
+- Compatible with WooCommerce version `9.8.1`.
+
+### Added
+
+- Added payment method used for order payment to meta on an order.
+- Added support for default WooCommerce terms and conditions checkbox on classic checkout.
+- Added support for hidden checkout fields on classic checkout.
+- Added improvements to admin settings panel.
+- Added fallback for configuration file fetch.
+- Added new logs for issue identification for widget and admin operations.
+- Added payment method title to orders.
+
+### Fixed
+
+- Fixed an issue in Classic checkout where the payment method would not load if merchant did not collect shipping address on their Checkout page.
+- Fixed an issue in Classic checkout when an order id failed to retrieve an order causing us not to navigate to order confirmation page post transaction approved.
+- Fixed an issue in Classic checkout in the sending of new order and processing state emails from WooCommerce.
+- Fixed overwriting of internal styling of button by external plugin.
+- Fixed checkout shipping form validation when toggling the "Ship to a different address?" checkbox with empty fields.
+- Fixed checkout overlay z-index to prevent switching payment methods during an active payment session.
+
 ## [1.3.1] - 2025-04-16
 
 ### Compatibility